wire lyrics POC into the animated player and fix sync (#113)
The POC modified NowPlaying.kt which isn't the active screen post-#106. Move
the lyrics icon into AnimatedPlayerSheet's bottom control row and collapse the
sheet on tap so the lyrics screen is actually visible.

Other fixes spotted while testing:

- LyricsScreen was computing positionMs as seekPosition*1000, but seekPosition
  is a 0..1 fraction, so the value passed to PositionChanged maxed at 1000ms
  and the highlight never advanced past line 0. Multiply by track.duration too.
- LyricsLineDto.time is Double. The server returns fractional ms like
  128021.99999999999 which Gson refused to parse into Long, killing every
  lyrics fetch.
- Drop the 300ms look-ahead in scheduleNextLine. Lines now advance at the
  actual line timestamp, which matches the audio for tracks with good LRC data.
- Replace deprecated Icons.Default.ArrowBack with the AutoMirrored variant.

+++ core/src/main/java/com/android/swingmusic/core/data/dto/LyricsLineDto.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class LyricsLineDto(
     @SerializedName("time")
-    val time: Long?,
+    val time: Double?,
     @SerializedName("text")
     val text: String?
 )
